+++ plugins/box/backup/languages/es.lang.php
@@ -1,17 +1,26 @@
 <?php
 
     return array(
-        'backup' => array(
-    		'Backups' => 'Backup',
-            'Backup date' => 'Backup dia',
-    		'Create Backup' => 'Crear Backup',
-    		'Delete' => 'Borrar',
+            'backup' => array(
+    	    'Backups' => 'Backups',
+	    'Backup' => 'Backup',
+            'Backup Date' => 'Fecha del Backup',
+    	    'Create Backup' => 'Crear Backup',
+	    'Restore' => 'Restaurar', 
+	    'Delete' => 'Eliminar',
             'storage' => 'Almacen',
-            'public' => 'publico',
+            'public' => 'público',
             'plugins' => 'plugins',
             'Size' => 'Tamaño',
             'Actions' => 'Acciones',
-            'Delete backup: :backup' => 'Borrar Backup: :backup',
-            'Creating...' => 'Creando...',
+            'Delete backup: :backup' => 'Eliminar Backup: :backup',
+            'Creating...' => 'Creando...', 
+            'Backup was created' => 'El Backup fue creado',
+            'Backup was not created' => 'El Backup no fue creado',
+            'Backup was deleted' => 'El Backup fue eliminado',
+            'Backup was not deleted' => 'El Backup no fue eliminado',
+            'Backup was restored' => 'El Backup fue restaurado',
+            'Unzip error' => 'Error de descompresión',
+            'Backup was not restored' => 'El Backup no fue restaurado',
         )
 	);