+++ plugins/box/dashboard/languages/es.lang.php
@@ -1,0 +1,11 @@
+<?php
+
+    return array(
+        'dashboard' => array(
+            'Dashboard' => 'Dashboard',
+            'Dashboard plugin for Monstra' => 'Dashboard plugin para Monstra',
+            'Welcome back' => 'Bienvenido nuevamente',
+            'Create New' => 'Crear Nuevo',
+            'Upload File' => 'Subir Archivo',
+        )
+    );